+++ rbc/runtime/include/rbc_graphics/frustum.h
@@ -5,12 +5,12 @@
 namespace rbc
 {
 RBC_RUNTIME_API bool frustum_cull(
-    luisa::float4x4 const& local_to_world,
+    luisa::double4x4 const& local_to_world,
     luisa::compute::AABB const& bounding,
-    luisa::span<luisa::float4, 6> frustum_planes,
-    luisa::float3 frustum_min_point,
-    luisa::float3 frustum_max_point,
-    luisa::float3 cam_forward,
-    luisa::float3 cam_pos
+    luisa::span<luisa::double4, 6> frustum_planes,
+    luisa::double3 frustum_min_point,
+    luisa::double3 frustum_max_point,
+    luisa::double3 cam_forward,
+    luisa::double3 cam_pos
 );
 }
